// react-pizza/src/redux/Slices/filter/types.ts
export enum SortPropertyEnum {
    RATING_DESC = 'rating',
    RATING_ASC = '-rating',
    TITLE_DESC = 'title',
    TITLE_ASC = '-title',
    PRICE_DESC = 'price',
    PRICE_ASC = '-price',
  }
  
  export type Sortpopup = {
    name: string;
    sort: SortPropertyEnum
  }
  
  export interface FilterSliceState {
    searchValue: string;
    categoryId: number,
    pageCount: number,
    sort: Sortpopup
  }